+++ test/integration/index.js
@@ -1,5 +1,6 @@
 const fs = require('fs')
 const path = require('path')
+const pump = require('pump')
 const browserify = require('browserify')
 const tests = fs.readdirSync(path.join(__dirname, 'lib'))
 const bundlePath = path.join(__dirname, 'bundle.js')
@@ -9,11 +10,17 @@ const b = browserify()
 const writeStream = fs.createWriteStream(bundlePath)
 
 tests.forEach(function (fileName) {
-  b.add(path.join(__dirname, 'lib', fileName))
+  const filePath = path.join(__dirname, 'lib', fileName)
+  console.log(`bundling test "${filePath}"`)
+  b.add(filePath)
 })
 
-b.bundle()
-.pipe(writeStream)
-.on('error', (err) => {
-  throw err
-})
+pump(
+  b.bundle(),
+  writeStream,
+  (err) => {
+    if (err) throw err
+    console.log(`Integration test build completed: "${bundlePath}"`)
+    process.exit(0)
+  }
+)